+++ block-storage/src/main/scala/coop/rchain/blockstorage/InMemBlockStore.scala
@@ -1,17 +1,14 @@
 package coop.rchain.blockstorage
 
+import scala.language.higherKinds
+
 import cats._
+import cats.effect.Sync
 import cats.effect.concurrent.Ref
-import cats.effect.{ExitCase, Sync}
 import cats.implicits._
 import coop.rchain.blockstorage.BlockStore.BlockHash
 import coop.rchain.casper.protocol.BlockMessage
 import coop.rchain.metrics.Metrics
-import coop.rchain.shared.SyncVarOps
-
-import scala.collection.mutable.ListBuffer
-import scala.concurrent.SyncVar
-import scala.language.higherKinds
 
 class InMemBlockStore[F[_], E] private ()(implicit
                                           monadF: Monad[F],
@@ -58,6 +55,7 @@ object InMemBlockStore {
 
   def createWithId: BlockStore[Id] = {
     import coop.rchain.metrics.Metrics.MetricsNOP
+    import coop.rchain.catscontrib.effect.implicits._
     val refId                         = emptyMapRef[Id](syncId)
     implicit val metrics: Metrics[Id] = new MetricsNOP[Id]()(syncId)
     InMemBlockStore.create(syncId, refId, metrics)
@@ -66,20 +64,4 @@ object InMemBlockStore {
   def emptyMapRef[F[_]](implicit syncEv: Sync[F]): F[Ref[F, Map[BlockHash, BlockMessage]]] =
     Ref[F].of(Map.empty[BlockHash, BlockMessage])
 
-  val syncId: Sync[Id] = new Sync[Id] {
-    def pure[A](x: A): cats.Id[A] = x
-
-    def handleErrorWith[A](fa: cats.Id[A])(f: Throwable => cats.Id[A]): cats.Id[A] = ???
-
-    def raiseError[A](e: Throwable): cats.Id[A] = ???
-
-    def bracketCase[A, B](acquire: cats.Id[A])(use: A => cats.Id[B])(
-        release: (A, cats.effect.ExitCase[Throwable]) => cats.Id[Unit]): cats.Id[B] = ???
-
-    def flatMap[A, B](fa: cats.Id[A])(f: A => cats.Id[B]): cats.Id[B] =
-      implicitly[Monad[Id]].flatMap(fa)(f)
-    def tailRecM[A, B](a: A)(f: A => cats.Id[Either[A, B]]): cats.Id[B] = ???
-
-    def suspend[A](thunk: => cats.Id[A]): cats.Id[A] = thunk
-  }
 }
